requirements.txt
torch
torchvision
torchaudio
pandas
numpy
scikit-learn
Pillow
